database migration with checks

+++ database/migrations/2016_05_12_215616_make_championship_default_game.php
@@ -12,9 +12,11 @@ class MakeChampionshipDefaultGame extends Migration
      */
     public function up()
     {
-        $query = "INSERT INTO `champ_games` (`name`, `description`, `uri`, `created_at`, `updated_at`)
+        if (Schema::connection('mysql_champ')->hasTable('games')) {
+            $query = "INSERT INTO `champ_games` (`name`, `description`, `uri`, `created_at`, `updated_at`)
             VALUES ('unknown', 'Unknown game', '', NULL, NULL);";
-        DB::connection('mysql_champ')->insert($query);
+            DB::connection('mysql_champ')->insert($query);
+        }
     }
 
     /**
@@ -24,6 +26,8 @@ class MakeChampionshipDefaultGame extends Migration
      */
     public function down()
     {
-        DB::connection('mysql_champ')->delete("DELETE FROM `champ_games` WHERE `name`='unknown'");
+        if (Schema::connection('mysql_champ')->hasTable('games')) {
+            DB::connection('mysql_champ')->delete("DELETE FROM `champ_games` WHERE `name`='unknown'");
+        }
     }
 }
